提交

+++ wine/src/com/chimu/wine/service/BannerService.java
@@ -26,13 +26,15 @@ public class BannerService {
     }
 
     public void modifyBanner(BannerBean bannerBean, MultipartFile file) throws Exception {
-        // 先删除原来的图片文件，再添加
-        String url = bannerBean.getImgurl().replace(Constant.Host, Constant.SaveImagesLocalPath);
-        FileGlobal.RemoveFile(url);
-        // 添加
-        String imageUrl = FileGlobal.AddFile(file);
-        if (CMString.isValid(imageUrl)) {
-            bannerBean.setImgurl(imageUrl);
+        if (file != null && !file.isEmpty()) {
+            // 先删除原来的图片文件，再添加
+            String url = bannerBean.getImgurl().replace(Constant.Host, Constant.SaveImagesLocalPath);
+            FileGlobal.RemoveFile(url);
+            // 添加
+            String imageUrl = FileGlobal.AddFile(file);
+            if (CMString.isValid(imageUrl)) {
+                bannerBean.setImgurl(imageUrl);
+            }
         }
         bannerDao.modifyBanner(bannerBean);
     }
